Keep historical v2 file for aggregator tests
