+++ tests/windows_aggregator_resize_v2_test.py
@@ -5,7 +5,8 @@ import os
 
 import nibabel as nib
 import tensorflow as tf
-
+import pandas as pd
+import numpy as np
 from niftynet.engine.sampler_resize_v2 import ResizeSampler
 from niftynet.engine.windows_aggregator_resize import ResizeSamplesAggregator
 from niftynet.io.image_reader import ImageReader
@@ -147,14 +148,144 @@ class ResizeSamplesAggregatorTest(tf.test.TestCase):
                 except tf.errors.OutOfRangeError:
                     break
                 more_batch = aggregator.decode_batch(
-                    out['image'], out['image_location'])
-        output_filename = '{}_niftynet_out.nii.gz'.format(
+                    {'window_image':out['image']}, out['image_location'])
+        output_filename = 'window_image_{}_niftynet_out.nii.gz'.format(
             sampler.reader.get_subject_id(0))
         output_file = os.path.join('testing_data',
                                    'aggregated',
                                    output_filename)
         self.assertAllClose(
             nib.load(output_file).shape, (256, 168, 256, 1, 2))
+        sampler.close_all()
+
+    def test_3d_init_mo(self):
+        reader = get_3d_reader()
+        sampler = ResizeSampler(reader=reader,
+                                window_sizes=MULTI_MOD_DATA,
+                                batch_size=1,
+                                shuffle=False,
+                                queue_length=50)
+        aggregator = ResizeSamplesAggregator(
+            image_reader=reader,
+            name='image',
+            output_path=os.path.join('testing_data', 'aggregated'),
+            interp_order=3)
+        more_batch = True
+
+        with self.test_session() as sess:
+            sampler.set_num_threads(2)
+            while more_batch:
+                try:
+                    out = sess.run(sampler.pop_batch_op())
+                except tf.errors.OutOfRangeError:
+                    break
+                sum_val = np.sum(out['image'])
+                more_batch = aggregator.decode_batch(
+                    {'window_image':out['image'], 'csv_sum':sum_val},
+                    out['image_location'])
+        output_filename = 'window_image_{}_niftynet_out.nii.gz'.format(
+            sampler.reader.get_subject_id(0))
+        sum_filename = os.path.join('testing_data','aggregated',
+                                    'csv_sum_{}_niftynet_out.csv'.format(
+            sampler.reader.get_subject_id(0)))
+        output_file = os.path.join('testing_data',
+                                   'aggregated',
+                                   output_filename)
+        self.assertAllClose(
+            nib.load(output_file).shape, (256, 168, 256, 1, 2))
+        sum_pd = pd.read_csv(sum_filename)
+        self.assertAllClose(
+            sum_pd.shape, [2, 2]
+        )
+        sampler.close_all()
+
+    def test_3d_init_mo_2im(self):
+        reader = get_3d_reader()
+        sampler = ResizeSampler(reader=reader,
+                                window_sizes=MULTI_MOD_DATA,
+                                batch_size=1,
+                                shuffle=False,
+                                queue_length=50)
+        aggregator = ResizeSamplesAggregator(
+            image_reader=reader,
+            name='image',
+            output_path=os.path.join('testing_data', 'aggregated'),
+            interp_order=3)
+        more_batch = True
+
+        with self.test_session() as sess:
+            sampler.set_num_threads(2)
+            while more_batch:
+                try:
+                    out = sess.run(sampler.pop_batch_op())
+                except tf.errors.OutOfRangeError:
+                    break
+                more_batch = aggregator.decode_batch(
+                    {'window_image':out['image'], 'window_im2':out['image']},
+                    out['image_location'])
+        output_filename = 'window_image_{}_niftynet_out.nii.gz'.format(
+            sampler.reader.get_subject_id(0))
+        outim2_filename = os.path.join('testing_data','aggregated',
+                                    'window_im2_{}_niftynet_out.nii.gz'.format(
+            sampler.reader.get_subject_id(0)))
+        output_file = os.path.join('testing_data',
+                                   'aggregated',
+                                   output_filename)
+        self.assertAllClose(
+            nib.load(output_file).shape, (256, 168, 256, 1, 2))
+        self.assertAllClose(
+            nib.load(outim2_filename).shape, (256, 168, 256, 1, 2))
+        sampler.close_all()
+
+    def test_3d_init_mo_3out(self):
+        reader = get_3d_reader()
+        sampler = ResizeSampler(reader=reader,
+                                window_sizes=MULTI_MOD_DATA,
+                                batch_size=1,
+                                shuffle=False,
+                                queue_length=50)
+        aggregator = ResizeSamplesAggregator(
+            image_reader=reader,
+            name='image',
+            output_path=os.path.join('testing_data', 'aggregated'),
+            interp_order=3)
+        more_batch = True
+
+        with self.test_session() as sess:
+            sampler.set_num_threads(2)
+            while more_batch:
+                try:
+                    out = sess.run(sampler.pop_batch_op())
+                except tf.errors.OutOfRangeError:
+                    break
+                sum_val = np.sum(out['image'])
+                stats_val = [np.sum(out['image']), np.min(out['image']),
+                                                         np.max(out['image'])]
+                more_batch = aggregator.decode_batch(
+                    {'window_image':out['image'], 'csv_sum':sum_val,
+                     'csv_stats':stats_val},
+                    out['image_location'])
+        output_filename = 'window_image_{}_niftynet_out.nii.gz'.format(
+            sampler.reader.get_subject_id(0))
+        sum_filename = os.path.join('testing_data','aggregated',
+                                    'csv_sum_{}_niftynet_out.csv'.format(
+            sampler.reader.get_subject_id(0)))
+        stats_filename = os.path.join('testing_data', 'aggregated',
+                                    'csv_stats_{}_niftynet_out.csv'.format(
+                                        sampler.reader.get_subject_id(0)))
+        output_file = os.path.join('testing_data',
+                                   'aggregated',
+                                   output_filename)
+        self.assertAllClose(
+            nib.load(output_file).shape, (256, 168, 256, 1, 2))
+        sum_pd = pd.read_csv(sum_filename)
+        self.assertAllClose(
+            sum_pd.shape, [2, 2]
+        )
+        stats_pd = pd.read_csv(stats_filename)
+        self.assertAllClose(
+            stats_pd.shape, [2, 4]
+        )
         sampler.close_all()
 
     def test_2d_init(self):
@@ -179,8 +310,8 @@ class ResizeSamplesAggregatorTest(tf.test.TestCase):
                 except tf.errors.OutOfRangeError:
                     break
                 more_batch = aggregator.decode_batch(
-                    out['image'], out['image_location'])
-        output_filename = '{}_niftynet_out.nii.gz'.format(
+                    {'window_image':out['image']}, out['image_location'])
+        output_filename = 'window_image_{}_niftynet_out.nii.gz'.format(
             sampler.reader.get_subject_id(0))
         output_file = os.path.join('testing_data',
                                    'aggregated',
@@ -211,8 +342,8 @@ class ResizeSamplesAggregatorTest(tf.test.TestCase):
                 except tf.errors.OutOfRangeError:
                     break
                 more_batch = aggregator.decode_batch(
-                    out['image'], out['image_location'])
-        output_filename = '{}_niftynet_out.nii.gz'.format(
+                    {'window_image':out['image']}, out['image_location'])
+        output_filename = 'window_image_{}_niftynet_out.nii.gz'.format(
             sampler.reader.get_subject_id(0))
         output_file = os.path.join('testing_data',
                                    'aggregated',
@@ -243,8 +374,8 @@ class ResizeSamplesAggregatorTest(tf.test.TestCase):
                 except tf.errors.OutOfRangeError:
                     break
                 more_batch = aggregator.decode_batch(
-                    out['label'], out['label_location'])
-        output_filename = '{}_niftynet_out.nii.gz'.format(
+                    {'window_label':out['label']}, out['label_location'])
+        output_filename = 'window_label_{}_niftynet_out.nii.gz'.format(
             sampler.reader.get_subject_id(0))
         output_file = os.path.join(
             'testing_data', 'aggregated', output_filename)
